pyproject.toml
[project]
name = "oreilly-ai-agents-gcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk[a2a,eval]>=1.16.0",
    "google-auth>=2.41.1",
    "google-cloud-bigquery>=3.38.0",
    "pytest>=8.4.2",
    "toolbox-core>=0.5.2",
    "yfinance>=0.2.66",
]
